+++ src/birdnet_tests/acoustic_models/v2_4/model_py/test_encode/test_acoustic_encode_custom_model.py
@@ -19,7 +19,7 @@ def test_custom_from_analyzer_v2_4_tf_fp32() -> None:
     "tf",
     get_model_path("acoustic", "2.4", "tf", "fp32"),
     get_lang_dir("acoustic", "2.4", "tf") / "en_us.txt",
-    library="tf",
+    library="tflite",
     precision="fp32",
     check_validity=False,
   )
@@ -37,7 +37,7 @@ def test_custom_from_analyzer_v2_4_tf_fp16() -> None:
     "tf",
     get_model_path("acoustic", "2.4", "tf", "fp16"),
     get_lang_dir("acoustic", "2.4", "tf") / "en_us.txt",
-    library="tf",
+    library="tflite",
     precision="fp16",
     check_validity=False,
   )
@@ -55,7 +55,7 @@ def test_custom_from_analyzer_v2_4_tf_int8() -> None:
     "tf",
     get_model_path("acoustic", "2.4", "tf", "int8"),
     get_lang_dir("acoustic", "2.4", "tf") / "en_us.txt",
-    library="tf",
+    library="tflite",
     precision="int8",
     check_validity=False,
   )